+++ part3/ch28/28-4-4-generic-not-equals/src/Vector2.h
@@ -11,7 +11,6 @@
 
 #include <initializer_list>
 #include <algorithm>
-#include <iostream>
 
 template<typename T>
 class Vector2_iter {
@@ -23,17 +22,14 @@ public:
 	T* operator->() { return cur; }
 
 	// 2016.10.10 add const
-	bool operator==(Vector2_iter other) const
-	{
-		return cur == other.cur;
-	}
+	bool operator==(Vector2_iter other) const { return cur == other.cur; }
 
 	Vector2_iter& operator++() { ++cur; return *this; }
 };
 
 template<typename T>
 class Vector2 {
-	int sz;
+	size_t sz;
 	T* elem;
 public:
 	using value_type = T;
